+++ posthog/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Next
 
+## 6.5.0 - 2026-02-16
+
+- feat: Support remote config for error tracking ([#415](https://github.com/PostHog/posthog-android/pull/415))
+
 ## 6.4.0 - 2026-02-05
 
 - fix: Filter out failed flags to prevent cached values from being overwritten during transient errors ([#408](https://github.com/PostHog/posthog-android/pull/408))